+++ Assembly-CSharp.csproj
@@ -65,12 +65,12 @@
      <Compile Include="Assets\Crayon\Scripts\Library\CrayonStateData.cs" />
      <Compile Include="Assets\Crayon\Scripts\Library\CrayonTweenCoroutines.cs" />
      <Compile Include="Assets\Crayon\Scripts\Library\CrayonUtils.cs" />
-     <Compile Include="Assets\Scripts\BufferReader.cs" />
-     <Compile Include="Assets\Scripts\SampleBuffer.cs" />
-     <Compile Include="Assets\Scripts\SensorEventBroadcaster.cs" />
-     <Compile Include="Assets\Scripts\SensorListener.cs" />
-     <Compile Include="Assets\Scripts\SensorSample.cs" />
-     <Compile Include="Assets\Scripts\UIListener.cs" />
+     <Compile Include="Assets\Scripts\Polling\BufferReader.cs" />
+     <Compile Include="Assets\Scripts\Polling\SampleBuffer.cs" />
+     <Compile Include="Assets\Scripts\Polling\SensorEventBroadcaster.cs" />
+     <Compile Include="Assets\Scripts\Polling\SensorListener.cs" />
+     <Compile Include="Assets\Scripts\Polling\SensorSample.cs" />
+     <Compile Include="Assets\Scripts\Polling\UIListener.cs" />
      <Compile Include="Assets\Scripts\Webhooks\SimpleHTTPServerComponent.cs" />
      <Compile Include="Assets\Scripts\Webhooks\WebhooksManager.cs" />
      <None Include="Assets\Crayon\UserPresets\inflate-1.txt" />
