--- Assets/Scripts/BufferReader.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 2f8885cc2a8ee4fa0be6ec5c40d6ed9c
-timeCreated: 1524184350
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 300
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
