+++ Assets/Scripts/Polling.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 7a80a55c16d1749a58d337336107c446
+folderAsset: yes
+timeCreated: 1525783925
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
