+++ Assets/Scripts/Polling/SensorSample.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d54896cd43a3642d8a093cfd292e6727
+timeCreated: 1524105965
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
