+++ Assets/Scripts/Polling/UIListener.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 00db030462480471b99ea8d32a719b7f
+timeCreated: 1523908302
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
