--- Assets/Scripts/SampleBuffer.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 183a23d81b3544dea8a9d73827023a42
-timeCreated: 1524184350
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 200
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
