--- Assets/Scripts/SensorEventBroadcaster.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 0feb5aa4d859b4a2f91a411a58fd8e1e
-timeCreated: 1524184351
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 400
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
