--- Assets/Scripts/SensorListener.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 18c5abd354f4b416b8b86b5ebd7639ec
-timeCreated: 1524184350
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 100
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
